Refactor 'getResponseForJsonwpError' to errors.js
