+++ lib/mjsonwp/errors.js
@@ -658,4 +658,31 @@ function getResponseForW3CError (err) {
   return [httpStatus, httpResBody];
 }
 
-export { MJSONWPError, errors, isErrorType, errorFromCode, getResponseForW3CError };
+function getResponseForJsonwpError (err) {
+  let httpStatus = 500;
+  let httpResBody = {
+    status: err.jsonwpCode,
+    value: {
+      message: err.message
+    }
+  };
+
+  if (isErrorType(err, errors.BadParametersError)) {
+    // respond with a 400 if we have bad parameters
+    log.debug(`Bad parameters: ${err}`);
+    httpStatus = 400;
+    httpResBody = err.message;
+  } else if (isErrorType(err, errors.NotYetImplementedError) ||
+             isErrorType(err, errors.NotImplementedError)) {
+    // respond with a 501 if the method is not implemented
+    httpStatus = 501;
+  } else if (isErrorType(err, errors.NoSuchDriverError)) {
+    // respond with a 404 if there is no driver for the session
+    httpStatus = 404;
+  }
+
+
+  return [httpStatus, httpResBody];
+}
+
+export { MJSONWPError, errors, isErrorType, errorFromCode, getResponseForW3CError, getResponseForJsonwpError };
