+++ lib/mjsonwp/mjsonwp.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { logger, util } from 'appium-support';
 import { validators } from './validators';
-import { errors, isErrorType, MJSONWPError, errorFromCode, getResponseForW3CError } from './errors';
+import { errors, isErrorType, MJSONWPError, errorFromCode, getResponseForW3CError, getResponseForJsonwpError } from './errors';
 import { METHOD_MAP, NO_SESSION_ID_COMMANDS } from './routes';
 import B from 'bluebird';
 
@@ -160,33 +160,6 @@ function makeArgs (requestParams, jsonObj, payloadParams, protocol) {
   // the list
   args = args.concat(urlParams.map((u) => requestParams[u]));
   return args;
-}
-
-function getResponseForJsonwpError (err) {
-  let httpStatus = 500;
-  let httpResBody = {
-    status: err.jsonwpCode,
-    value: {
-      message: err.message
-    }
-  };
-
-  if (isErrorType(err, errors.BadParametersError)) {
-    // respond with a 400 if we have bad parameters
-    log.debug(`Bad parameters: ${err}`);
-    httpStatus = 400;
-    httpResBody = err.message;
-  } else if (isErrorType(err, errors.NotYetImplementedError) ||
-             isErrorType(err, errors.NotImplementedError)) {
-    // respond with a 501 if the method is not implemented
-    httpStatus = 501;
-  } else if (isErrorType(err, errors.NoSuchDriverError)) {
-    // respond with a 404 if there is no driver for the session
-    httpStatus = 404;
-  }
-
-
-  return [httpStatus, httpResBody];
 }
 
 function routeConfiguringFunction (driver) {
